+++ internal/delivery/http/request/exchange_request/exchange.go
@@ -32,12 +32,12 @@ type UpdateWithdrawalSetting struct {
 } // @name UpdateWithdrawalSetting
 
 type GetWithdrawalsRequest struct {
-	CurrencyID *string `json:"currency_id,omitempty" query:"currency_id"`
-	DateFrom   *string `json:"date_from" query:"date_from"`
-	DateTo     *string `json:"date_to" query:"date_to"`
-	Slug       *string `json:"slug,omitempty" query:"slug"`
-	Page       *uint32 `json:"page,omitempty" query:"page"`
-	PageSize   *uint32 `json:"page_size,omitempty" query:"page_size"`
+	CurrencyID *string `json:"currency_id,omitempty" query:"currency_id" validate:"omitempty"`
+	DateFrom   *string `json:"date_from" query:"date_from" validate:"omitempty,datetime=2006-01-02T15:04:05Z07:00"`
+	DateTo     *string `json:"date_to" query:"date_to" validate:"omitempty,datetime=2006-01-02T15:04:05Z07:00,gtfield=DateFrom"`
+	Slug       *string `json:"slug,omitempty" query:"slug" validate:"omitempty,alphanum"`
+	Page       *uint32 `json:"page,omitempty" query:"page" validate:"omitempty,min=1"`
+	PageSize   *uint32 `json:"page_size,omitempty" query:"page_size" validate:"omitempty,min=1,max=100"`
 } // @name ExchangeGetWithdrawalsRequest
 
 type GetWithdrawalsExportedRequest struct {